First attempt at switching pointForCoordinate:zoomLevel:tileSize: over to using floats for zoomLevel

+++ Sources/MRProjection.h
@@ -11,7 +11,7 @@
 
 // Convert pixel coordinates to latitude / longitude and vice-versa
 
-- (CGPoint)pointForCoordinate:(MRMapCoordinate)coordinate zoomLevel:(NSUInteger)zoom tileSize:(CGSize)tileSize;
+- (CGPoint)pointForCoordinate:(MRMapCoordinate)coordinate zoomLevel:(float)zoom tileSize:(CGSize)tileSize;
 - (MRMapCoordinate)coordinateForPoint:(CGPoint)point zoomLevel:(NSUInteger)zoom tileSize:(CGSize)tileSize;
 
 @end
